✨ F08: ADMIN > Create <AdminContent /> to add product
